\name{CfEstimateQuantiles-package}
\alias{CfEstimateQuantiles-package}
\alias{CfEstimateQuantiles}
\docType{package}
\title{Estimate quantiles using any order Cornish-Fisher expansion}
\description{Estimate quantiles using formula (18) from http://www.jaschke-net.de/papers/CoFi.pdf (Yaschke; 2001)
}
\details{
\tabular{ll}{
Package: \tab CfEstimateQuantiles\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-05-12\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}
}
\author{
Maxim Yurchuk <maxim.yurchuk@gmail.com>
}
\references{
http://www.jaschke-net.de/papers/CoFi.pdf
}
